Move the CBMR basis and optimizer numerics into the cbmr package
Phase 1, part two. Two blocks of CBMR-only code lived in shared modules:

- nimare/utils.py held coef_spline_bases and b_spline_bases, 123 lines with a
  single caller and no other CBMR code around them -> cbmr/basis.py.
- nimare/meta/utils.py held the scipy L-BFGS-B backend, 191 lines interleaved
  with the unrelated ALE and KDA kernel machinery -> cbmr/optimizers.py.

Moved outright rather than re-exported from the old locations. A re-export would
mean nimare.utils importing nimare.meta.cbmr, whose package __init__ imports the
estimator, which imports nimare.utils -- a cycle. Neither name appears in
docs/api.rst and each had one real caller, so there is nothing to keep
compatible.

Dropping the moved block left scipy.optimize unused in nimare/meta/utils.py.

Also fixes a bug in the previous commit: the rewritten cleanup in
test_meta_package_defers_cbmr_import deleted every nimare.meta.* module absent
from its saved set, which permanently removed nimare.meta.utils and
nimare.meta.cbma.* and left later tests importing fresh duplicates. That
surfaced as 13 unrelated ALE failures, including a PicklingError, since pickling
resolves functions by qualified name and the module objects no longer matched. It
now snapshots every nimare.meta module up front and restores that mapping
exactly.

196 CBMR, ALE and utils tests pass. A broad run over nimare/tests leaves one
pre-existing failure, test_corr_transform_performance[...kda...], confirmed to
fail identically with these changes stashed.
